hasAttribute() check
