+++ packages/phpstan-rules/src/Rules/RequireInvokableControllerRule.php
@@ -6,10 +6,8 @@ namespace Symplify\PHPStanRules\Rules;
 
 use Nette\Utils\Strings;
 use PhpParser\Node;
-use PhpParser\Node\AttributeGroup;
 use PhpParser\Node\Stmt\ClassMethod;
 use PHPStan\Analyser\Scope;
-use Symplify\Astral\Naming\SimpleNameResolver;
 use Symplify\PHPStanRules\ValueObject\MethodName;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
@@ -23,21 +21,6 @@ final class RequireInvokableControllerRule extends AbstractInvokableControllerRu
      * @var string
      */
     public const ERROR_MESSAGE = 'Use invokable controller with __invoke() method instead';
-
-    /**
-     * @var string
-     */
-    private const ROUTE_ATTRIBUTE = 'Symfony\Component\Routing\Annotation\Route';
-
-    /**
-     * @var SimpleNameResolver
-     */
-    private $simpleNameResolver;
-
-    public function __construct(SimpleNameResolver $simpleNameResolver)
-    {
-        $this->simpleNameResolver = $simpleNameResolver;
-    }
 
     /**
      * @return string[]
@@ -120,20 +103,5 @@ CODE_SAMPLE
         }
 
         return $this->hasAttribute($node);
-    }
-
-    private function hasAttribute(ClassMethod $node): bool
-    {
-        /** @var AttributeGroup $attrGroup */
-        foreach ((array) $node->attrGroups as $attrGroup) {
-            foreach ($attrGroup->attrs as $attribute) {
-                $attributeClass = $this->simpleNameResolver->getName($attribute->name);
-                if ($attributeClass === self::ROUTE_ATTRIBUTE) {
-                    return true;
-                }
-            }
-        }
-
-        return false;
     }
 }
